Final changes
It appears that the webpack changes I made in the vue.config.js file were the root cause of the issue.

I do get warnings that certain files are too large, but I believe these are to be ignored.  Im unsure if the webpack configuration would be beneficial or is just a warning remover.