Added discord chat and changed a few visuals
Added a discord.svg logo in case I want to use it in the future.

Footer has been fixed to display a coffee icon and also link to Ping's direct repo on github.

Updated the dashboard to include showing discord links.  These are found in the `chat_url` part of the api call.  I filter for something matching `discord` and then display that link.  Otherwise, no link is displayed.